+++ src/parsers/mod.rs
@@ -3,6 +3,7 @@ pub mod drawing_xml;
 pub mod dynamic_attr_records;
 pub mod general_xml;
 pub mod jproperties;
+pub mod magic;
 pub mod string_scan;
 pub mod xml_util;